Rename kodion.utils.datetime_parser to kodion.utils.datetime
- Module includes a number of non-parsing related datetime utility methods
